perform/ml/requirements.txt
xgboost>=3.0
scikit-learn>=1.5
pandas>=2.0
nfl-data-py>=0.3
joblib>=1.4
fastapi>=0.110
uvicorn>=0.30
psycopg2-binary>=2.9
python-dotenv>=1.0
scipy>=1.12
numpy>=2.0
